support writing files greater than 2Gb
individual buffers need to be less than 2Gb. There are still "bare" ints
at several levels, e.g. arguments to functions in iozero.c and possibly
in mpi itself.

+++ libsw/mpmy_gather.c
@@ -177,8 +177,8 @@ MPMY_NGather(const void *sendbuf, int count, MPMY_Datatype type,
     int chan;
     int doc;
     int sendproc;
-    int bufsz;
-    int inbytes;
+    long bufsz;
+    long inbytes;
     unsigned int mask;
     void *buf;
     int procnum = MPMY_Procnum();
